new frn. no alpha

+++ models/layers.py
@@ -72,8 +72,8 @@ class NewFilterResponseNormalization(nn.Module):
 
             # A = torch.max(self.limit, alpha * a +
             #               torch.abs(self.eps))
-            part1 = a > self.limit
-            part2 = a <= self.limit
+            part1 = (a > self.limit).float()
+            part2 = (a <= self.limit).float()
             A = part1*a/self.limit + part2*self.limit
             x = x / torch.sqrt(A + 1e-6)
             x = torch.max(self.gamma * x + self.beta, self.tau)
